code refactored

+++ contracts/aggregator/fantom/flashloan/helpers.sol
@@ -1,7 +1,7 @@
 //SPDX-License-Identifier: Unlicense
 pragma solidity ^0.8.0;
-import "./variables.sol";
-import "../../common/helpers.sol";
+import './variables.sol';
+import '../../common/helpers.sol';
 
 contract Helper is HelpersCommon, Variables {
     using SafeERC20 for IERC20;
@@ -12,12 +12,14 @@ contract Helper is HelpersCommon, Variables {
      * @param token_ token for which allowance is to be given.
      * @param spender_ the address to which the allowance is to be given.
      * @param amount_ amount of token.
-    */
-    function approve(address token_, address spender_, uint256 amount_) internal {
+     */
+    function approve(
+        address token_,
+        address spender_,
+        uint256 amount_
+    ) internal {
         TokenInterface tokenContract_ = TokenInterface(token_);
-        try tokenContract_.approve(spender_, amount_) {
-            
-        } catch {
+        try tokenContract_.approve(spender_, amount_) {} catch {
             IERC20 token = IERC20(token_);
             token.safeApprove(spender_, 0);
             token.safeApprove(spender_, amount_);
@@ -30,16 +32,16 @@ contract Helper is HelpersCommon, Variables {
      * @param _instaLoanVariables struct which includes list of token addresses and amounts.
      * @param _fees list of premiums/fees for the corresponding addresses for flashloan.
      * @param _receiver address to which tokens have to be approved.
-    */
+     */
     function safeApprove(
         FlashloanVariables memory _instaLoanVariables,
         uint256[] memory _fees,
         address _receiver
     ) internal {
         uint256 length_ = _instaLoanVariables._tokens.length;
-        require(length_ == _instaLoanVariables._amounts.length, "Lengths of parameters not same");
-        require(length_ == _fees.length, "Lengths of parameters not same");
-        for (uint i = 0; i < length_; i++) {
+        require(length_ == _instaLoanVariables._amounts.length, 'Lengths of parameters not same');
+        require(length_ == _fees.length, 'Lengths of parameters not same');
+        for (uint256 i = 0; i < length_; i++) {
             approve(_instaLoanVariables._tokens[i], _receiver, _instaLoanVariables._amounts[i] + _fees[i]);
         }
     }
@@ -49,14 +51,11 @@ contract Helper is HelpersCommon, Variables {
      * @notice Transfers the tokens to the receiver address.
      * @param _instaLoanVariables struct which includes list of token addresses and amounts.
      * @param _receiver address to which tokens have to be transferred.
-    */
-    function safeTransfer(
-        FlashloanVariables memory _instaLoanVariables,
-        address _receiver
-    ) internal {
+     */
+    function safeTransfer(FlashloanVariables memory _instaLoanVariables, address _receiver) internal {
         uint256 length_ = _instaLoanVariables._tokens.length;
-        require(length_ == _instaLoanVariables._amounts.length, "Lengths of parameters not same");
-        for (uint i = 0; i < length_; i++) {
+        require(length_ == _instaLoanVariables._amounts.length, 'Lengths of parameters not same');
+        for (uint256 i = 0; i < length_; i++) {
             IERC20 token = IERC20(_instaLoanVariables._tokens[i]);
             token.safeTransfer(_receiver, _instaLoanVariables._amounts[i]);
         }
@@ -66,12 +65,13 @@ contract Helper is HelpersCommon, Variables {
      * @dev Validates if the receiver sent the correct amounts of funds.
      * @notice Validates if the receiver sent the correct amounts of funds.
      * @param _instaLoanVariables struct which includes list of initial balances, final balances and fees for the respective tokens.
-    */
-    function validateFlashloan(
-        FlashloanVariables memory _instaLoanVariables
-    ) internal pure {
-        for (uint i = 0; i < _instaLoanVariables._iniBals.length; i++) {
-            require(_instaLoanVariables._iniBals[i] + _instaLoanVariables._instaFees[i] <= _instaLoanVariables._finBals[i], "amount-paid-less");
+     */
+    function validateFlashloan(FlashloanVariables memory _instaLoanVariables) internal pure {
+        for (uint256 i = 0; i < _instaLoanVariables._iniBals.length; i++) {
+            require(
+                _instaLoanVariables._iniBals[i] + _instaLoanVariables._instaFees[i] <= _instaLoanVariables._finBals[i],
+                'amount-paid-less'
+            );
         }
     }
 
@@ -79,16 +79,16 @@ contract Helper is HelpersCommon, Variables {
      * @dev Returns fee for the passed route in BPS.
      * @notice Returns fee for the passed route in BPS. 1 BPS == 0.01%.
      * @param _route route number for flashloan.
-    */
-    function calculateFeeBPS(uint256 _route) public view returns(uint256 BPS_){
+     */
+    function calculateFeeBPS(uint256 _route) public view returns (uint256 BPS_) {
         if (_route == 9) {
             BPS_ = aaveV3Lending.FLASHLOAN_PREMIUM_TOTAL();
-        } else if (_route == 10){
+        } else if (_route == 10) {
             BPS_ = InstaFeeBPS;
         } else {
-            revert("Invalid source");
+            revert('Invalid source');
         }
-        
+
         if (BPS_ < InstaFeeBPS) {
             BPS_ = InstaFeeBPS;
         }
@@ -98,7 +98,7 @@ contract Helper is HelpersCommon, Variables {
      * @dev Returns to true if the passed address is a DSA else returns false.
      * @notice Returns to true if the passed address is a DSA else returns false.
      * @param _account account to check for, if DSA.
-    */
+     */
     function checkIfDsa(address _account) internal view returns (bool) {
         return instaList.accountID(_account) > 0;
     }
@@ -107,11 +107,11 @@ contract Helper is HelpersCommon, Variables {
      * @dev  better checking by double encoding the data.
      * @notice better checking by double encoding the data.
      * @param data_ data passed.
-    */
+     */
     modifier verifyDataHash(bytes memory data_) {
         bytes32 dataHash_ = keccak256(data_);
-        require(dataHash_ == dataHash && dataHash_ != bytes32(0), "invalid-data-hash");
-        require(status == 2, "already-entered");
+        require(dataHash_ == dataHash && dataHash_ != bytes32(0), 'invalid-data-hash');
+        require(status == 2, 'already-entered');
         dataHash = bytes32(0);
         _;
         status = 1;
@@ -120,11 +120,11 @@ contract Helper is HelpersCommon, Variables {
     /**
      * @dev reentrancy gaurd.
      * @notice reentrancy gaurd.
-    */
-    modifier reentrancy {
-        require(status == 1, "already-entered");
+     */
+    modifier reentrancy() {
+        require(status == 1, 'already-entered');
         status = 2;
         _;
-        require(status == 1, "already-entered");
+        require(status == 1, 'already-entered');
     }
 }
