+++ contracts/aggregator/fantom/flashloan/variables.sol
@@ -1,11 +1,10 @@
 //SPDX-License-Identifier: Unlicense
 pragma solidity ^0.8.0;
 
-import "./interfaces.sol";
-import "../../common/interface.sol";
+import './interfaces.sol';
+import '../../common/interface.sol';
 
 contract ConstantVariables {
-
     address public constant aaveV3LendingAddr = 0x794a61358D6845594F94dc1DB02A252b5b4814aD;
     IAaveV3Lending public constant aaveV3Lending = IAaveV3Lending(aaveV3LendingAddr);
 
@@ -14,18 +13,16 @@ contract ConstantVariables {
     ListInterface public constant instaList = ListInterface(instaListAddr);
 
     uint256 public constant InstaFeeBPS = 5; // in BPS; 1 BPS = 0.01%
-
 }
 
 contract Variables is ConstantVariables {
-
     address public owner;
 
     uint256 internal ownerStatus;
 
     bytes32 internal dataHash;
     // if 1 then can enter flashlaon, if 2 then callback
-    uint internal status;
+    uint256 internal status;
 
     struct FlashloanVariables {
         address[] _tokens;
@@ -35,8 +32,8 @@ contract Variables is ConstantVariables {
         uint256[] _instaFees;
     }
 
-    mapping (uint256 => address) public routeToImpl;
-    mapping (uint256 => bool) public routeStatus;
+    mapping(uint256 => address) public routeToImpl;
+    mapping(uint256 => bool) public routeStatus;
     address internal implToCall;
     uint256[] public routes;
 }
